- fix warnings on Mac
git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathtext@44306 27541ba8-7e3a-0410-8455-c3a389f83636

+++ core/mathtext/src/fontembed.cc
@@ -257,7 +257,7 @@ namespace mathtext {
 			return font_data;
 		}
 		if(fread(&font_data[0], sizeof(uint8_t),
-				 length, fp) != static_cast<int>(length)) {
+				 length, fp) != static_cast<unsigned long>(length)) {
 			perror("fread");
 			font_data.clear();
 			return font_data;
